Integrate the new "nanokernel" PLL from Dave Mills.
This code is backwards compatible with the older "microkernel" PLL, but
allows ntpd v4 to use nanosecond resolution.  Many other improvements.

PPS_SYNC and hardpps() are NOT supported yet.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@44574 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/sys/time.h
@@ -31,7 +31,7 @@
  * SUCH DAMAGE.
  *
  *	@(#)time.h	8.5 (Berkeley) 5/4/95
- * $Id: time.h,v 1.34 1998/10/23 10:42:42 phk Exp $
+ * $Id: time.h,v 1.35 1998/12/15 17:38:32 des Exp $
  */
 
 #ifndef _SYS_TIME_H_
@@ -141,7 +141,7 @@ struct timecounter {
 	char			*tc_name;
 	void			*tc_priv;
 	/* These fields will be managed by the generic code. */
-	int32_t			tc_adjustment;
+	int64_t			tc_adjustment;
 	u_int32_t		tc_scale_micro;
 	u_int32_t		tc_scale_nano_i;
 	u_int32_t		tc_scale_nano_f;
